Fixed slider metrics & adding important render features.

+++ src/draw/draw.cpp
@@ -56,7 +56,7 @@ void ekg::draw::rect(float x, float y, float w, float h, const ekg::vec4 &color,
     allocator.dispatch();
 }
 
-void ekg::draw::sync_scissor_pos(int32_t x, int32_t y) {
+void ekg::draw::sync_scissor_pos(float x, float y) {
     ekg::core->get_gpu_allocator().sync_scissor_pos(x, y);
 }
